backend/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LangChain/LangGraph (>= 1.0 for robustness)
langchain>=1.0.0
langchain-aws>=1.0.0
langgraph>=1.0.0
langchain-community>=1.0.0
langchain-core>=1.0.0
langsmith>=0.1.0  # For evaluation and agent testing

# AWS
boto3>=1.40.0
botocore>=1.40.0

# Data generation
Faker>=37.0.0
sdv[copula]>=1.26.0
pandas>=2.3.0
numpy>=2.0.0

# PII detection with spacy
spacy>=3.7.0

# Document generation
python-docx>=1.1.0
pypdf2>=3.0.1
reportlab>=4.2.0
openpyxl>=3.1.2

# Streaming
sse-starlette>=2.1.0

# OpenAPI parsing
openapi-spec-validator>=0.7.1
prance>=23.6.0

# Utilities
python-dotenv>=1.0.0  # Load .env files
python-multipart  # File uploads
aiofiles  # Async file ops
redis>=5.0.0  # For caching
httpx>=0.27.0  # Async HTTP client

# Development tools
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.6.0

